fix: resolve golangci-lint issues
Fixed 21 linter issues:
- godoclint (7): Updated godoc comments to start with symbol names
- revive unused-parameter (13): Renamed unused params to underscore
- gocognit (1): Refactored ScanWiFi to reduce cognitive complexity from 25 to <20

Refactoring details:
- Extracted buildSSIDLookup and buildBSSIDLookup helper functions
- Extracted checkRogueAPs, checkChannelInterference, checkUtilization methods
- Reduced ScanWiFi main function complexity while maintaining functionality

Build verified: go build ./... passes
All 45 files committed with lint fixes applied.

+++ internal/sap/link/link_test.go
@@ -826,7 +826,7 @@ func TestMonitorDefaultValues(t *testing.T) {
 	m := link.NewMonitor("eth0")
 
 	// Check poll interval
-	expectedInterval := int64(link.DefaultPollInterval.Milliseconds())
+	expectedInterval := link.DefaultPollInterval.Milliseconds()
 	if m.MonitorPollInterval() != expectedInterval {
 		t.Errorf("expected poll interval %d ms, got %d ms", expectedInterval, m.MonitorPollInterval())
 	}
